requirements.txt
python-dotenv
openai
flask
python-dotenv
